+++ index.html
@@ -229,5 +229,7 @@
     </div>
   </template>
   <script src="./js/main.js"></script>
+  <script src="./js/data.js"></script>
+  <script src="./js/util.js"></script>
 </body>
 </html>
